refactor: 카드 타입을 추정하는 로직 분리
